tests: don't run named pipes tests on windows
