+++ all_test.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"strings"
 	"testing"
-	"syscall"
 
 	. "github.com/otiai10/mint"
 )
@@ -16,14 +15,6 @@ func TestMain(m *testing.M) {
 	code := m.Run()
 	teardown(m)
 	os.Exit(code)
-}
-
-func setup(m *testing.M) {
-	os.MkdirAll("test/data.copy", os.ModePerm)
-	os.Symlink("test/data/case01", "test/data/case03/case01")
-	os.Chmod("test/data/case07/dir_0555", 0555)
-	os.Chmod("test/data/case07/file_0444", 0444)
-	syscall.Mkfifo("test/data/case11/foo/bar", 0555)
 }
 
 func teardown(m *testing.M) {
@@ -89,27 +80,7 @@ func TestCopy(t *testing.T) {
 		Expect(t, err).ToBe(nil)
 	})
 
-	When(t, "specified src contains a folder with a named pipe", func(t *testing.T) {
-		dest:= "test/data.copy/case11"
-		err := Copy("test/data/case11", dest)
-		Expect(t, err).ToBe(nil)
-
-		info, err := os.Lstat("test/data/case11/foo/bar")
-		Expect(t, err).ToBe(nil)
-		Expect(t, info.Mode()&os.ModeNamedPipe != 0).ToBe(true)
-		Expect(t, info.Mode().Perm()).ToBe(os.FileMode(0555))
-	})
-
-	When(t, "specified src is a named pipe", func(t *testing.T) {
-		dest:= "test/data.copy/case11/foo/bar.named"
-		err := Copy("test/data/case11/foo/bar", dest)
-		Expect(t, err).ToBe(nil)
-
-		info, err := os.Lstat(dest)
-		Expect(t, err).ToBe(nil)
-		Expect(t, info.Mode()&os.ModeNamedPipe != 0).ToBe(true)
-		Expect(t, info.Mode().Perm()).ToBe(os.FileMode(0555))
-	})
+	testPipes(t)
 }
 
 func TestOptions_OnSymlink(t *testing.T) {
